// src/presentation/contexts/book-form/types.ts
export type BookForm = {
    title: {
        isWrongFill: boolean
        text: string
    },
    author: {
        isWrongFill: boolean
        text: string
    },
    pages: {
        isWrongFill: boolean
        text: string
    },
    description: {
        isWrongFill: boolean
        text: string
    },
    currentPage: {
        isWrongFill: boolean
        text: string
    },
}